+++ Ship.h
@@ -62,7 +62,7 @@ public:
 	// Get this ship's description.
 	const std::string &Description() const;
 	// Get this ship's cost.
-	int Cost() const;
+	int64_t Cost() const;
 	// Get the licenses needed to buy or operate this ship.
 	const std::vector<std::string> &Licenses(const Government *government) const;
